+++ internal/parser/parser.go
@@ -1,4 +1,4 @@
-package main
+package parser
 
 import (
 	"bufio"
@@ -6,22 +6,17 @@ import (
 	"os"
 	"regexp"
 	"strings"
+
+	"github.com/matthewrobinsdev/kindle-notes-parser/pkg/models"
 )
 
-type Highlight struct {
-	Title    string
-	Author   string
-	Page     string
-	Location string
-	Date     string
-	Text     string
-}
-
-func parseClippings(filename string) ([]Highlight, error) {
+func ParseClippings(filename string) ([]models.Highlight, error) {
 	file, err := os.Open(filename)
+
 	if err != nil {
 		return nil, err
 	}
+
 	defer file.Close()
 
 	var lines []string
@@ -33,9 +28,9 @@ func parseClippings(filename string) ([]Highlight, error) {
 	return createHighlights(lines), nil
 }
 
-func createHighlights(lines []string) []Highlight {
-	var highlights []Highlight
-	var currentHighlight Highlight
+func createHighlights(lines []string) []models.Highlight {
+	var highlights []models.Highlight
+	var currentHighlight models.Highlight
 
 	title := regexp.MustCompile(`^(.*) \((.*)\)$`)
 	var utf8BOM = []byte{0xEF, 0xBB, 0xBF}
@@ -67,9 +62,31 @@ func createHighlights(lines []string) []Highlight {
 		if len(line) > 0 && !strings.HasPrefix(line, "==========") {
 			currentHighlight.Text = line
 			highlights = append(highlights, currentHighlight)
-			currentHighlight = Highlight{}
+			currentHighlight = models.Highlight{}
 		}
 	}
 
 	return highlights
+}
+
+func GroupHighlightsByBook(highlights []models.Highlight) []models.BookGroup {
+	bookMap := make(map[string][]models.Highlight)
+	authorMap := make(map[string]string)
+
+	for _, highlight := range highlights {
+		bookMap[highlight.Title] = append(bookMap[highlight.Title], highlight)
+		authorMap[highlight.Title] = highlight.Author
+	}
+
+	var books []models.BookGroup
+	for title, highlights := range bookMap {
+		books = append(books, models.BookGroup{
+			Title:      title,
+			Author:     authorMap[title],
+			Highlights: highlights,
+			Expanded:   false,
+		})
+	}
+
+	return books
 }
