+++ pkg/models/highlight.go
@@ -1,0 +1,23 @@
+package models
+
+type Highlight struct {
+	Title    string
+	Author   string
+	Page     string
+	Location string
+	Date     string
+	Text     string
+}
+
+type BookGroup struct {
+	Title      string
+	Author     string
+	Highlights []Highlight
+	Expanded   bool
+}
+
+type ListItem struct {
+	IsBook         bool
+	BookIndex      int
+	HighlightIndex int
+}
